# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: 公论 (GongLun) API
  description: 公论结构化讨论系统后端 API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 本地开发服务器
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /topics/{topicId}/relations:
    get:
      tags: [关系]
      summary: 获取话题关系列表
      parameters:
        - in: path
          name: topicId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 关系列表

    post:
      tags: [关系]
      summary: 创建消息关系
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: topicId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [relationType, sourceMessageId, targetRefs]
              properties:
                relationType:
                  type: string
                  enum: [QUOTE, REPLY, SUPPORT, OPPOSE, CORRECT, LINK, UNLINK]
                sourceMessageId:
                  type: string
                targetRefs:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [targetMessageId]
                    properties:
                      targetMessageId:
                        type: string
                      targetSelectedText:
                        type: string
                      targetSelectedTextHash:
                        type: string
      responses:
        '201':
          description: 关系创建成功
        '403':
          description: 话题已归档
        '404':
          description: 消息不存在